add "exports" to package.json
when using

    import { ModelView } from 'toad.js'

rollup includes all of toad.
(see https://rollupjs.org/guide/en/#tree-shaking-doesnt-seem-to-be-working)

with the "exports" directive one can now write

  import { ModelView } from 'toad.js/view/ModelView'

instead of

  import { ModelView } from 'toad.js/lib/src/view/ModelView'

the "typesVersions" is there for vscode and rollup.

but while the import isn't marked as wrong in vscode, completion when
writing the import directive doesn't work.  yet.